refactor: Move tests into a separate file to simplify code navigation (#7126)
